move invoke into Interpreter
